+++ packages/repo-app-import-sub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/repo-app-download-sub-generator
 
+## 1.0.16
+
+### Patch Changes
+
+- @sap-ux/fiori-elements-writer@3.0.16
+- @sap-ux/odata-service-inquirer@3.0.9
+
 ## 1.0.15
 
 ### Patch Changes